fix: use compute for post data serilization
