+++ lib/app/features/feed/create_post/views/pages/post_form_modal/hooks/use_post_quill_controller.dart
@@ -7,10 +7,12 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_quill/flutter_quill.dart';
 import 'package:flutter_quill/quill_delta.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:ion/app/extensions/delta.dart';
 import 'package:ion/app/features/feed/data/models/entities/modifiable_post_data.f.dart';
+import 'package:ion/app/features/feed/providers/parsed_media_provider.r.dart';
 import 'package:ion/app/features/ion_connect/model/event_reference.f.dart';
+import 'package:ion/app/features/ion_connect/model/media_attachment.dart';
 import 'package:ion/app/features/ion_connect/providers/ion_connect_entity_provider.r.dart';
-import 'package:ion/app/features/ion_connect/views/hooks/use_parsed_media_content.dart';
 
 QuillController? usePostQuillController(
   WidgetRef ref, {
@@ -38,7 +40,12 @@ QuillController? usePostQuillController(
       }
       if (modifiedEntity != null) {
         if (modifiedEntity is ModifiablePostEntity) {
-          final (:content, :media) = parseMediaContent(data: modifiedEntity.data);
+          final (:content, :media) =
+              ref.watch(parsedMediaProvider(modifiedEntity.data)).valueOrNull ??
+                  (
+                    content: Delta().blank,
+                    media: <MediaAttachment>[],
+                  );
           final document = Document.fromDelta(content);
           return QuillController(
             document: document,
